+++ src/types/events.ts
@@ -1,0 +1,5 @@
+export interface WalletProviderEvent {
+  type: 'error' | 'warn' | 'info';
+  title: string;
+  message: string;
+}
