+++ src/types/theme.ts
@@ -1,0 +1,7 @@
+export interface ModalTheme {
+  background: string;
+  main: string;
+  secondary: string;
+  border: string;
+  hover: string;
+}
